rag_mcp_server/requirements.txt
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
rank_bm25
python-dotenv
